+++ frontend/app/archive/page.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from "react";
 import Link from "next/link";
-import { supabase } from "@/utils/supabaseClient";
+import { supabase } from "@/lib/supabase";
 import type { Session } from "@supabase/supabase-js";
 
 interface PollInfo {
